+++ CHANGELOG.md
@@ -178,6 +178,10 @@ music.loop = true
 - added `onVirtualButtonPress()`, `onVirtualButtonDown()`, `onVirtualButtonRelease()`
 - fixed touches not treated as mouse
 - (**BREAK**) changed `onTouchStart()`, `onTouchMove()` and `onTouchEnd()` callback signature to `(pos: Vec2, touch: Touch) => void` (exposes the native `Touch` object)
+- added `onGamepadButtonPress()`, `onGamepadButtonDown()`, `onGamepadButtonRelease()`
+- added `isGamepadButtonPressed()`, `isGamepadButtonDown()`, `isGamepadButtonRelease()`
+- added `getConnectedGamepads()`
+- added `onGamepadConnect()` and `onGamepadDisconnect()`
 
 ## Level